+++ out/codecoverage/codeCoverageHtml.js
@@ -14,17 +14,18 @@ const fs = require("fs");
 const path = require("path");
 const codeCoverage_1 = require("./codeCoverage");
 class CodeCoverageHtml {
-    static getHtmlForCoverage(codeCoverage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter) {
+    static getHtmlForCoverage(codeCoverage, isSmall, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter) {
         return __awaiter(this, void 0, void 0, function* () {
             if (!codeCoverage) {
                 return '';
             }
             let content = `
-		<table class="table table-striped">
+		<table class="${!isSmall ? "table table-striped" : ""}">
 			<thead>
 				<tr>
 					<th style="width: 40%">Apex Class</th>
 					<th style="width: 60%">Coverage</th>
+					${!isSmall ? '<th>&nbsp;</th>' : ''}
 				</tr>
 			</thead>
 			<tbody>`;
@@ -34,7 +35,7 @@ class CodeCoverageHtml {
                 }).forEach((item) => {
                     const className = item.name;
                     const percentage = item.coveredPercent;
-                    const contentItem = CodeCoverageHtml.calculateHtmlForItem(className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter);
+                    const contentItem = CodeCoverageHtml.calculateHtmlForItem(codeCoverage, isSmall, className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter);
                     if (!contentItem) {
                         return;
                     }
@@ -47,7 +48,7 @@ class CodeCoverageHtml {
                 }).forEach((item) => {
                     const className = item.name;
                     const percentage = item.percentage ? Number.parseFloat(item.percentage.replace('%', '')) : 0;
-                    const contentItem = CodeCoverageHtml.calculateHtmlForItem(className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter);
+                    const contentItem = CodeCoverageHtml.calculateHtmlForItem(codeCoverage, isSmall, className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter);
                     if (!contentItem) {
                         return;
                     }
@@ -75,8 +76,14 @@ class CodeCoverageHtml {
                 if (found === undefined) {
                     content += `
 					<tr>
-						<td><span class="apexClassName">${apexClass}</span></td><td><div class="progress"><div class="progress-bar bg-danger" role="progressbar" style="width: 100%;" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">No coverage</div></div></td>
+						<td><span class="apexClassName">${apexClass}</span></td>
+						<td><div class="progress"><div class="progress-bar bg-danger" role="progressbar" style="width: 100%;" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">No coverage</div></div></td>
+						${!isSmall ? '<td>&nbsp;</td>' : ''}
 					</tr>`;
+                    if (!isSmall) {
+                        content += `
+					<tr class="collapsible"><td colspan="3">&nbsp;</td></tr>`;
+                    }
                 }
             });
             const triggers = yield fsPromises.readdir(codeCoverage_1.apexTriggersDirPath);
@@ -94,8 +101,14 @@ class CodeCoverageHtml {
                 if (found === undefined) {
                     content += `
 					<tr>
-						<td><span class="apexClassName">${apexClass}</span></td><td><div class="progress"><div class="progress-bar bg-danger" role="progressbar" style="width: 100%;" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">No coverage</div></div></td>
+						<td><span class="apexClassName">${apexClass}</span></td>
+						<td><div class="progress"><div class="progress-bar bg-danger" role="progressbar" style="width: 100%;" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">No coverage</div></div></td>
+						${!isSmall ? '<td>&nbsp;</td>' : ''}
 					</tr>`;
+                    if (!isSmall) {
+                        content += `
+					<tr class="collapsible"><td colspan="3">&nbsp;</td></tr>`;
+                    }
                 }
             });
             content += `
@@ -104,7 +117,7 @@ class CodeCoverageHtml {
             return content;
         });
     }
-    static calculateHtmlForItem(className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter) {
+    static calculateHtmlForItem(codeCoverage, isSmall, className, percentage, lowCoverageFilter, projectFilesOnlyFilter, fileNameFilter) {
         if (!className) {
             return null;
         }
@@ -130,10 +143,33 @@ class CodeCoverageHtml {
             colorClass = "bg-success";
         }
         let coverage = Math.round(percentage).toString();
-        const content = `
-			<tr>
-				<td><span class="apexClassName">${className}</span></td><td><div class="progress"><div class="progress-bar ${colorClass}" role="progressbar" style="width: ${coverage}%;" aria-valuenow="${coverage}" aria-valuemin="0" aria-valuemax="100">${coverage}%</div></div></td>
+        let testClasses = codeCoverage_1.CodeCoverage.getTestClassesForApexClass(className, codeCoverage);
+        const hasTestInfo = testClasses !== null && testClasses.length > 0;
+        let content = `
+			<tr id="${className}">
+				<td><span class="apexClassName">${className}</span></td>
+				<td><div class="progress"><div class="progress-bar ${colorClass}" role="progressbar" style="width: ${coverage}%;" aria-valuenow="${coverage}" aria-valuemin="0" aria-valuemax="100">${coverage}%</div></div></td>
+				${!isSmall && hasTestInfo ? '<td><i class="fas fa-chevron-circle-down" title="Show test class coverage contribution" onclick="showInfo(this)"></i></td>' : ""}</td>
+				${!isSmall && !hasTestInfo ? '<td>&nbsp;</td>' : ''}
 			</tr>`;
+        if (!isSmall) {
+            if (hasTestInfo) {
+                let testInfo = '';
+                testClasses === null || testClasses === void 0 ? void 0 : testClasses.forEach(test => {
+                    testInfo += `${test.apexClassOrTriggerName}.${test.apexTestMethodName}: ${test.percentage}<br />`;
+                });
+                content += `
+				<tr id="${className}--test" class="collapsible">
+					<td>&nbsp;</td>
+					<td>${testInfo}</td>
+					<td>&nbsp;</td>
+				</tr>`;
+            }
+            else {
+                content += `
+				<tr class="collapsible"><td colspan="3">&nbsp;</td></tr>`;
+            }
+        }
         return content;
     }
 }