+++ out/codecoverage/index.js
@@ -1,8 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.CodeCoveragePanel = exports.CodeCoverage = void 0;
+exports.CodeCoverageSideViewPanelProvider = exports.CodeCoveragePanel = exports.CodeCoverage = void 0;
 var codeCoverage_1 = require("./codeCoverage");
 Object.defineProperty(exports, "CodeCoverage", { enumerable: true, get: function () { return codeCoverage_1.CodeCoverage; } });
 var codeCoverageViewPanel_1 = require("./codeCoverageViewPanel");
 Object.defineProperty(exports, "CodeCoveragePanel", { enumerable: true, get: function () { return codeCoverageViewPanel_1.CodeCoveragePanel; } });
+var codeCoverageSideViewPanel_1 = require("./codeCoverageSideViewPanel");
+Object.defineProperty(exports, "CodeCoverageSideViewPanelProvider", { enumerable: true, get: function () { return codeCoverageSideViewPanel_1.CodeCoverageSideViewPanelProvider; } });
 //# sourceMappingURL=index.js.map