+++ src/codecoverage/index.ts
@@ -1,2 +1,3 @@
 export { CodeCoverage } from './codeCoverage';
 export { CodeCoveragePanel } from './codeCoverageViewPanel';
+export { CodeCoverageSideViewPanelProvider } from './codeCoverageSideViewPanel';